requirements.txt
asgiref==3.5.1
Django==4.0.4
sqlparse==0.4.2
tzdata==2022.1
Pillow==9.1.1
django-ckeditor~=6.4.2